Scrollbar bug fix, refactoring
Scrollbars disappear and reappear now more reliably.
However, in double page mode, it is still(!) possible to make the viewport too
small for the image but not small enough for the horizontal scrollbar to appear.
Furthermore, it is still(!) possible to make scrollbars appear when they are not
necessary (both single page mode and double page mode).

Changes in MainWindow:
- interface of _show_scrollbars changed
- _show_scrollbars simplified
- _hscroll replaced with _scroll[constants.WIDTH_AXIS]
- _vscroll replaced with _scroll[constants.HEIGHT_AXIS]

Changes in zoom.py:
- docs added for _smaller


git-svn-id: svn+ssh://svn.code.sf.net/p/mcomix/code@974 ef8490bc-f317-4725-81ed-e95a9aca1bcf